düzeltme

+++ Dortislem/Matematik.cs
@@ -19,7 +19,7 @@ namespace Dortislem
 
         public void Bolme (int sayi1, int sayi2)
         {
-            float sonuc = sayi1 / sayi2;
+            int sonuc = sayi1 / sayi2;
             Console.WriteLine("sonuc : " + sonuc);
         }
